Move OpenAi entity to service module
