+++ core-entities/src/commonMain/kotlin/org/pointyware/artes/entities/hosts/Host.kt
@@ -3,19 +3,7 @@ package org.pointyware.artes.entities.hosts
 /**
  *
  */
-sealed interface Host {
+interface Host {
     val id: Long
     val title: String
-}
-
-/**
- *
- */
-class OpenAi(
-    override val id: Long,
-    override val title: String,
-    val orgId: String,
-    val apiKey: String,
-): Host {
-
 }
